fix: change weight from int to float

+++ backend/app/models/sessions.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, Float
 from datetime import datetime, timezone
 from sqlalchemy.orm import relationship
 from ..db import Base
@@ -40,7 +40,7 @@ class SessionCriterion(Base):
     session_id = Column(Integer, ForeignKey("sessions.id"), primary_key=True)
     criterion_id = Column(Integer, ForeignKey("criteria.id"), primary_key=True)
     role_id = Column(Integer, ForeignKey("roles.id"), primary_key=True)
-    weight = Column(Integer, nullable=False, default=1)
+    weight = Column(Float, nullable=False, default=1)
 
     session = relationship("Session", back_populates="session_criteria_assoc")
     criterion = relationship("Criterion", back_populates="session_criteria_assoc")
